move 'store' config to 'store' folder
